Land_Registory_Contract_Project.sol
Land registry is a time taking and complicated process. A lot of intermediaries are involved in this process. This project makes the complex and time-consuming process of Land registry more accelerant, more secure and a simple and easy system.

This Project explains the possibilities of a more efficient Land registry system that is Etherium Blockchain-based. Following are the main objectives of this project
To protect the owner’s rights
To prevent sale Fraud
To resolve disputes
To make sure ownership is correctly transferred to a new owner
To remove intermediaries eg property seller